+++ src/examples/transfer_funds.js
@@ -19,6 +19,10 @@ async function transferFunds() {
     assetId: Coinbase.assets.Eth,
     destination: anotherWallet,
   });
+
+  // Wait for the transfer to complete or fail on-chain.
+  await transfer.wait();
+
   console.log(`Transfer successfully completed: `, transfer.toString());
 }